Add Weather Station Graph
A project I created when I was still new to VBA and programming. I found this tool helpful in tracking weather changes as I am sensitive to large fluctuations in barometric pressure and temperature. This could probably be recreated with HTTP requests but I haven't the time to work on anything leisurely at the moment.